- refactoring to more explicitly use macroscopic network and macroscopic link segments where appropriate + preparing to better support removing of ids gaps through (network)builders

+++ src/main/java/org/planit/utils/network/virtual/ConnectoidSegment.java
@@ -17,6 +17,6 @@ public interface ConnectoidSegment extends EdgeSegment {
    * 
    * @return connectoid segment id
    */
-  int getConnectoidSegmentId();
+  long getConnectoidSegmentId();
 
 }
